Fix pylint findings
