+++ pyproject.toml
@@ -49,7 +49,7 @@ cli.env = { AWS_PROFILE = "kha-restricted", KHA_DATA_S3_BUCKET = "kha-store-dev"
 cli.help = "Run the command line interface"
 deploy.cmd = "zappa deploy prod"
 deploy.help = "Initially deploy to kommtheuteaktenzeichen.de"
-linter.cmd = "pylint --enable-all-extensions kha"
+linter.cmd = "pylint --enable-all-extensions kha tests"
 linter.help = "Check for style violations"
 scraper.script = "scripts.scraper:run"
 scraper.help = "Scrape from web interface"
